requirements.txt
phidata
openai
duckduckgo-search
yfinance
fastapi[standard]
sqlalchemy
lancedb 
tantivy 
pypdf 
sqlalchemy
duckdb
python-dotenv